Clean up skills settings structure
