+++ crates/ironclaw_webui_v2_static/static/js/i18n/en.js
@@ -450,14 +450,11 @@ registerPack("en", {
     "Skills extend the agent with domain-specific instructions. Add a SKILL.md bundle or place SKILL.md files in your workspace.",
   "skills.failedLoad": "Failed to load skills: {message}",
   "skills.import": "Add skill",
-  "skills.importDesc": "Install from an HTTPS SKILL.md URL or paste SKILL.md content directly.",
+  "skills.importDesc": "Paste SKILL.md content to add a user-mounted skill.",
   "skills.name": "Skill name",
   "skills.namePlaceholder": "skill-name",
-  "skills.url": "HTTPS URL",
-  "skills.urlHint": "Use a direct HTTPS link to SKILL.md or a supported skill bundle.",
-  "skills.urlPlaceholder": "https://example.com/SKILL.md",
   "skills.content": "SKILL.md content",
-  "skills.contentHint": "Use this instead of URL when importing local or copied skill content.",
+  "skills.contentHint": "Use the full SKILL.md frontmatter and prompt content.",
   "skills.contentPlaceholder": "---\\nname: example\\ndescription: ...\\n---\\n",
   "skills.install": "Add",
   "skills.installing": "Adding...",
@@ -465,8 +462,6 @@ registerPack("en", {
   "skills.installedSuccess": "Added skill \"{name}\"",
   "skills.nameRequired": "Skill name is required.",
   "skills.contentRequired": "SKILL.md content is required.",
-  "skills.importSourceRequired": "Provide an HTTPS URL or SKILL.md content.",
-  "skills.httpsRequired": "URL must use HTTPS.",
   "skills.remove": "Remove",
   "skills.delete": "Delete",
   "skills.edit": "Edit",
